new changes

+++ src/main/java/com/example/saidabot/Repository/Level1_Repository.java
@@ -3,6 +3,6 @@ package com.example.saidabot.Repository;
 import com.example.saidabot.Model.CustomerCare100.Level1;
 import org.springframework.data.repository.CrudRepository;
 
-public interface Level1_Repository extends CrudRepository<Level1, Integer> {
+public interface Level1_Repository extends CrudRepository<Level1, Double> {
 
 }
